feat: netcdf chunked catchment variable reads
The feature enables reading chunks of netcdf variable values instead of
reading all values for a given variable. A 'hint_shared_provider_id'
method is added that enables 'hinting' to a provider one of the id's
that it must provide. this method can be repeatedly called to add
necessary catchments. the netcdf provider keeps a set of the id's that
have been hinted and rebuilds it's chunking index on the first timestep
which is used to read only the necessary data.

I am not satisfied with the hinting approach, but it gets the job done
for now. Alternative approaches are:

- pass down list of catchments when constructing a provider. This would
  require a lot of refactoring and extra fiddling that is currently not
  required.
  - or, create the shared forcing provider from all the id's and pass
    that down. has the same issues as described.
- include the catchment-id to the factory forcing provider. This would
  enable this functionality but transparently.